requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.0.6
django-cors-headers==4.4.0
django-redis==5.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.15.2
redis==5.0.7
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
